+++ src/main/java/com/mageddo/dnsproxyserver/config/entrypoint/ConfigEnv.java
@@ -43,6 +43,8 @@ public class ConfigEnv {
 
   private static final String MG_NO_REMOTE_SERVERS = "MG_NO_REMOTE_SERVERS";
 
+  private static final String MG_NO_ENTRIES_RESPONSE_CODE = "MG_NO_ENTRIES_RESPONSE_CODE";
+
   private Path currentPath;
   private String resolvConfPath;
   private String logFile;
@@ -55,6 +57,7 @@ public class ConfigEnv {
   private URI dockerHost;
   private Boolean resolvConfOverrideNameServers;
   private Boolean noRemoteServers;
+  private Integer noEntriesResponseCode;
 
   public static ConfigEnv fromEnv() {
     return ConfigEnv
@@ -71,6 +74,7 @@ public class ConfigEnv {
       .dockerHost(UriUtils.createURI(Envs.getStringOrNull(MG_DOCKER_HOST)))
       .resolvConfOverrideNameServers(Envs.getBooleanOrNull(MG_RESOLVCONF_OVERRIDE_NAMESERVERS))
       .noRemoteServers(Envs.getBooleanOrNull(MG_NO_REMOTE_SERVERS))
+      .noEntriesResponseCode(Envs.getIntegerOrNull(MG_NO_ENTRIES_RESPONSE_CODE))
       .build();
   }